+++ scripts/prepare-smoke-plugins.sh
@@ -1,0 +1,16 @@
+#!/bin/bash
+
+# For npm (Mutates packages/cli/package.json)
+npm run pack:prepare
+
+cp ./redocly-cli.tgz ./openapi-core.tgz ./__tests__/smoke-plugins
+
+echo "Current directory:"
+pwd
+echo
+echo "Current directory content:"
+ls -a
+echo
+echo "Target directory content:"
+ls -a __tests__/smoke-plugins/
+echo